refactor(schedule-import): split service file and address review feedback
- Break the oversized scheduleImportService.ts into a scheduleImport/
  folder: types.ts, parseCsv.ts, buildCommitPayload.ts, api.ts.
- Validate diff-schedule / commit-schedule responses with zod schemas
  (the DiffResult/CommitResult types are now inferred from them)
  instead of unchecked `as` casts.
- parseScheduleCsv now throws on any PapaParse error, including
  delimiter detection failures.
- Narrow resolveSetStageName to take just the stage name and move it
  below buildCommitPayload's return.
- Update component imports to the new paths.
